+++ src/client/components/ac_twice_passwords/ac_twice_passwords.js
@@ -10,7 +10,7 @@
  *  - AC: the acUserLogin internal data structure
  *      Is undefined when invoked from ac_reset_pwd template
  *      Take care!
- *  - ahName: the AccountsHub.ahClass instance name (passed from reset_ask through URL parameters)
+ *  - acName: the AccountsCore.acAccount instance name (passed from reset_ask through URL parameters)
  *      set from ac_reset_pwd (so exclusive from AC above)
  *  - role: 'signup|change|reset'
  *      This happens to also be the prefix of the to-be-called AccountsUI options methods
@@ -108,7 +108,7 @@ Template.ac_twice_passwords_sub.helpers({
                 ? this.AC.options.coloredBorders() === AccountsUI.C.Colored.MANDATORY : AccountsUI.opts().coloredBorders() === AccountsUI.C.Colored.MANDATORY;
         return {
             AC: this.AC,
-            ahName: this.ahName,
+            acName: this.acName,
             label: this.label || pwixI18n.label( I18N, 'twice_passwords.label' ),
             placeholder: this.placeholder1 || pwixI18n.label( I18N, 'twice_passwords.placeholder1' ),
             wantsLength: true,
@@ -126,7 +126,7 @@ Template.ac_twice_passwords_sub.helpers({
                 ? this.AC.options.coloredBorders() === AccountsUI.C.Colored.MANDATORY : AccountsUI.opts().coloredBorders() === AccountsUI.C.Colored.MANDATORY;
         return {
             AC: this.AC,
-            ahName: this.ahName,
+            acName: this.acName,
             label: '',
             placeholder: this.placeholder2 || pwixI18n.label( I18N, 'twice_passwords.placeholder2' ),
             withMandatoryBorder: mandatoryBorder