+++ src/server/js/methods.js
@@ -5,7 +5,7 @@
 const assert = require( 'assert' ).strict; // up to nodejs v16.x
 
 import { Accounts } from 'meteor/accounts-base';
-import { AccountsHub } from 'meteor/pwix:accounts-hub';
+import { AccountsCore } from 'meteor/pwix:accounts-core';
 import { Logger } from 'meteor/pwix:logger';
 
 const logger = Logger.get();
@@ -14,17 +14,17 @@ Meteor.methods({
     // All AccountsUI.byXxxx methods return a user object without the crypted password nor the profile
 
     // find the user who holds the given reset password token
-    async 'pwix.AccountsUI.m.byResetToken'( ahName, token ){
-        const ahInstance = AccountsHub.getInstance( ahName );
-        assert( ahInstance && ahInstance instanceof AccountsHub.ahClass, 'expects an instance of AccountsHub.ahClass, got '+ahInstance );
-        const doc = await ahInstance.collection().findOneAsync({ 'services.password.reset.token': token },{ 'services.password.reset': 1 });
-        return AccountsHub.s.cleanupUserDocument( doc );
+    async 'pwix.AccountsUI.m.byResetToken'( acName, token ){
+        const acInstance = AccountsCore.getInstance( acName );
+        assert( acInstance && acInstance instanceof AccountsCore.acAccount, 'expects an instance of AccountsCore.acAccount, got '+acInstance );
+        const doc = await acInstance.collection().findOneAsync({ 'services.password.reset.token': token },{ 'services.password.reset': 1 });
+        return AccountsCore.s.cleanupUserDocument( doc );
     },
 
     // find the user who holds the given email verification token
     async 'pwix.AccountsUI.m.byEmailVerificationToken'( token ){
         const doc = await Meteor.users.findOneAsync({ 'services.email.verificationTokens': { $elemMatch: { token: token }}},{ 'services.email':1 });
-        return AccountsHub.s.cleanupUserDocument( doc );
+        return AccountsCore.s.cleanupUserDocument( doc );
     },
 
     // create a user without auto login
@@ -42,16 +42,16 @@ Meteor.methods({
     // see https://github.com/meteor/meteor/blob/devel/packages/accounts-password/password_server.js#L529
     // do not send extra data when using the standard 'users' collection
     // return true|false
-    async 'pwix.AccountsUI.m.forgotPassword'( ahName, email ){
-        const ahInstance = AccountsHub.getInstance( ahName );
-        assert( ahInstance && ahInstance instanceof AccountsHub.ahClass, 'expects an instance of AccountsHub.ahClass, got '+ahInstance );
+    async 'pwix.AccountsUI.m.forgotPassword'( acName, email ){
+        const acInstance = AccountsCore.getInstance( acName );
+        assert( acInstance && acInstance instanceof AccountsCore.acAccount, 'expects an instance of AccountsCore.acAccount, got '+acInstance );
         let res = null;
-        const user = await ahInstance.byEmailAddress( email );
+        const user = await AccountsCore.byEmailAddress( acInstance, email );
         if( user ){
-            if( ahName === AccountsHub.ahOptions._defaults.name ){
+            if( acName === AccountsCore.ahOptions._defaults.name ){
                 res = await Accounts.sendResetPasswordEmail( user._id, email );
             } else {
-                res = await Accounts.sendResetPasswordEmail( user._id, email, undefined, { ahName: ahName });
+                res = await Accounts.sendResetPasswordEmail( user._id, email, undefined, { acName: acName });
             }
         }
         return Boolean( res );